don't allow selection of time or control elements
We disable the hidden toggle on click feature whenever there is a
selection on the page (so that you can copy a block of text without it
hiding). However, clicking on the `.control` element to toggle the
hidden state sometimes triggers a selection which makes subsequent
clicks not toggle the hidden state. Therefore, just disable selection in
the `.control` and child elements.