add swipe to reveal delete option
mimics ios7 behavior using the hammer.js library. swipe left on a note
to reveal a delete button, swipe right to hide it.

hammer.js is used to listen to swipeleft and swiperight events which
toggles the swiped class on a note view. the sliding in and out of the
delete button is done using css transforms and translations.