switch to ember.prod.js when building for deployment
Ember prod gets rid of debug statements and assertion calls, so it
should run a little faster. Unfortunately, there isn't a way to specify
which ember version to build using browserify because of lack of
conditionals. The solution here is a bit hacky; we copy either the prod
or dev version to ember-browserify.js and have browserify serve that
file as the ember module.