add checkerboard and striped background patterns to notes
Add a BackgroundGenerator class and assign an instance of it to each
note. A new instance will randomly choose a color using the
`randomcolor` library and a random pattern, either striped or
checkerboard.

The instance exposes the `toCSS()` method that returns a string of the
css rules for the background pattern.